feat: distribution of blocks by components in documents section (2022-10-12)
