+++ src/components/document/components/block-description/index.ts
@@ -1,0 +1,1 @@
+export { default } from './block-description.component.vue';
