+++ src/components/document/components/block-type-document/image/activities.png
