+++ src/components/document/components/block-type-document/image/case.png
