+++ src/components/document/components/block-type-document/image/responsibilities.png
